+++ src/main/java/baseball/utility/StringUtils.java
@@ -1,4 +1,4 @@
-package baseball.utils;
+package baseball.utility;
 
 import java.util.List;
