+++ src/test/java/baseball/balls/BallValueEqualsTest.java
@@ -1,5 +1,6 @@
 package baseball.balls;
 
+import baseball.domain.balls.BallValue;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
