+++ src/test/java/baseball/balls/BallsHasSameBallTest.java
@@ -1,5 +1,7 @@
 package baseball.balls;
 
+import baseball.domain.balls.Ball;
+import baseball.domain.balls.Balls;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
