+++ src/test/java/baseball/results/ResultsUtilToStringTest.java
@@ -1,6 +1,7 @@
 package baseball.results;
 
-import baseball.utils.ResultsUtils;
+import baseball.domain.results.ResultStatus;
+import baseball.utility.ResultsUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
