+++ src/test/java/baseball/utility/StringUtilsTest.java
@@ -1,4 +1,4 @@
-package baseball.utils;
+package baseball.utility;
 
 import java.util.List;
 import java.util.stream.Collectors;
