+++ src/test/java/baseball/validator/IntegerRangeValidateTest.java
@@ -1,4 +1,4 @@
-package baseball.validators;
+package baseball.validator;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
