+++ src/test/java/baseball/validator/ListValidatorsTest.java
@@ -1,4 +1,4 @@
-package baseball.validators;
+package baseball.validator;
 
 import java.util.List;
 import java.util.stream.IntStream;
